Add files via upload
Here are the complete details on how to build the physical robot that corresponds with the exciting integrated model. I have also included a link to the Tinkercad design for additional design information, as well as a link for circuit simulation, where you can find more information about the circuit used in the physical model.